+++ app/api/v1/healthcheck.rb
@@ -1,8 +1,14 @@
 class V1::Healthcheck < Grape::API
-  desc "Api healthcheck endpoint returns 'OK'" do
+  content_type :txt, "text/plain"
+
+  desc "Api healthcheck endpoint checks db connection and returns 'Healthy' or 'Unhealthy'" do
     tags ["miscellaneous"]
   end
+
   get :healthcheck do
-    {status: "OK"}
+    content_type "text/plain"
+    body "Healthy" if ActiveRecord::Base.connected?
+
+    "Unhealthy"
   end
 end